+++ src/main/java/com/societegenerale/commons/plugin/maven/MavenScopePathProvider.java
@@ -2,16 +2,23 @@ package com.societegenerale.commons.plugin.maven;
 
 import com.societegenerale.commons.plugin.model.RootClassFolder;
 import com.societegenerale.commons.plugin.service.ScopePathProvider;
+import org.apache.maven.project.MavenProject;
 
 public class MavenScopePathProvider implements ScopePathProvider {
 
+    private final MavenProject project;
+
+    public MavenScopePathProvider(MavenProject project) {
+        this.project = project;
+    }
+
     @Override
     public RootClassFolder getMainClassesPath() {
-        return new RootClassFolder("target/classes");
+        return new RootClassFolder(project.getBuild().getOutputDirectory());
     }
 
     @Override
     public RootClassFolder getTestClassesPath() {
-        return new RootClassFolder("target/test-classes");
+        return new RootClassFolder(project.getBuild().getTestOutputDirectory());
     }
 }